The DatabaseManager class is implemented to manage database connections and cache transfers. Deprecated functions have been removed from bot.py , added method calls to the new class for working with the cache. Updated documentation in README.md to reflect changes in architecture and functionality.
